refactor(alpen-ee): simplify state diff readiness check and isolate DA tests
