+++ functional-tests-new/entry.py
@@ -273,14 +273,6 @@ def main(argv: list[str]) -> int:
             mesh_bootnodes=True,
             enable_l1_da=True,
         ),
-        # Alpen-client with DA pipeline (sequencer-only, no fullnodes)
-        # Using batch_sealing_block_count=30 to accumulate enough state diff for
-        # multi-chunk DA testing (60 contracts @ ~3 per block = 20 blocks)
-        "alpen_client_da": AlpenClientEnv(
-            fullnode_count=0,
-            enable_l1_da=True,
-            batch_sealing_block_count=30,
-        ),
     }
 
     # Set up test runtime
